requirements.txt
streamlit>=1.25
numpy>=1.23
pillow>=9.5
opencv-python-headless>=4.7.0.72
matplotlib>=3.7